LPS-66426 Move tests to the right package
