+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/type/radio/internal/RadioDDMFormFieldValueAccessorTest.java
@@ -12,14 +12,12 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.select;
+package com.liferay.dynamic.data.mapping.type.radio.internal;
 
 import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
 import com.liferay.portal.json.JSONFactoryImpl;
-import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.util.LocaleUtil;
 
 import org.junit.Assert;
@@ -28,38 +26,23 @@ import org.junit.Test;
 /**
  * @author Renato Rego
  */
-public class SelectDDMFormFieldValueAccessorTest {
+public class RadioDDMFormFieldValueAccessorTest {
 
 	@Test
-	public void testGetSelectValue() throws Exception {
-		JSONArray expectedJSONArray = createJSONArray("value 1");
-
+	public void testGetRadioValue() {
 		DDMFormFieldValue ddmFormFieldValue =
 			DDMFormValuesTestUtil.createDDMFormFieldValue(
-				"Select", new UnlocalizedValue(expectedJSONArray.toString()));
+				"Radio", new UnlocalizedValue("[\"value 1\"]"));
 
-		SelectDDMFormFieldValueAccessor selectDDMFormFieldValueAccessor =
-			new SelectDDMFormFieldValueAccessor();
+		RadioDDMFormFieldValueAccessor radioDDMFormFieldValueAccessor =
+			new RadioDDMFormFieldValueAccessor();
 
-		selectDDMFormFieldValueAccessor.jsonFactory = _jsonFactory;
-
-		JSONArray actualJSONArray = selectDDMFormFieldValueAccessor.getValue(
-			ddmFormFieldValue, LocaleUtil.US);
+		radioDDMFormFieldValueAccessor.jsonFactory = new JSONFactoryImpl();
 
 		Assert.assertEquals(
-			expectedJSONArray.toString(), actualJSONArray.toString());
+			"value 1",
+			radioDDMFormFieldValueAccessor.getValue(
+				ddmFormFieldValue, LocaleUtil.US));
 	}
-
-	protected JSONArray createJSONArray(String... strings) {
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
-
-		for (String string : strings) {
-			jsonArray.put(string);
-		}
-
-		return jsonArray;
-	}
-
-	private final JSONFactory _jsonFactory = new JSONFactoryImpl();
 
 }
