+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/type/select/internal/SelectDDMFormFieldValueAccessorTest.java
@@ -12,12 +12,14 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.radio;
+package com.liferay.dynamic.data.mapping.type.select.internal;
 
 import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
 import com.liferay.portal.json.JSONFactoryImpl;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.util.LocaleUtil;
 
 import org.junit.Assert;
@@ -26,23 +28,38 @@ import org.junit.Test;
 /**
  * @author Renato Rego
  */
-public class RadioDDMFormFieldValueAccessorTest {
+public class SelectDDMFormFieldValueAccessorTest {
 
 	@Test
-	public void testGetRadioValue() {
+	public void testGetSelectValue() throws Exception {
+		JSONArray expectedJSONArray = createJSONArray("value 1");
+
 		DDMFormFieldValue ddmFormFieldValue =
 			DDMFormValuesTestUtil.createDDMFormFieldValue(
-				"Radio", new UnlocalizedValue("[\"value 1\"]"));
+				"Select", new UnlocalizedValue(expectedJSONArray.toString()));
 
-		RadioDDMFormFieldValueAccessor radioDDMFormFieldValueAccessor =
-			new RadioDDMFormFieldValueAccessor();
+		SelectDDMFormFieldValueAccessor selectDDMFormFieldValueAccessor =
+			new SelectDDMFormFieldValueAccessor();
 
-		radioDDMFormFieldValueAccessor.jsonFactory = new JSONFactoryImpl();
+		selectDDMFormFieldValueAccessor.jsonFactory = _jsonFactory;
+
+		JSONArray actualJSONArray = selectDDMFormFieldValueAccessor.getValue(
+			ddmFormFieldValue, LocaleUtil.US);
 
 		Assert.assertEquals(
-			"value 1",
-			radioDDMFormFieldValueAccessor.getValue(
-				ddmFormFieldValue, LocaleUtil.US));
+			expectedJSONArray.toString(), actualJSONArray.toString());
 	}
+
+	protected JSONArray createJSONArray(String... strings) {
+		JSONArray jsonArray = _jsonFactory.createJSONArray();
+
+		for (String string : strings) {
+			jsonArray.put(string);
+		}
+
+		return jsonArray;
+	}
+
+	private final JSONFactory _jsonFactory = new JSONFactoryImpl();
 
 }
